Removed CUD stuff; simplified package layout as a result.
