+++ django_webauth/__init__.py
@@ -2,4 +2,4 @@ __version__ = '0.3.2'
 
 # Here so that the user detail provisioning signal handlers
 # definitely get hooked up.
-from . import sources
+from . import ldap
